+++ src/movements/bishop.rs
@@ -1,0 +1,8 @@
+use crate::errs::Errs;
+use crate::piece::Board;
+
+fn bishop<'a>(pieces: &Board, strs: &Vec<char>, is_white: bool) -> Result<Board, Errs> {
+    // todo 動こうとしたマスが直線上にあるか
+    // todo 動こうとしたマスがなにかの駒で防がれていないか
+    todo!()
+}
