+++ src/movements/king.rs
@@ -1,0 +1,8 @@
+use crate::errs::Errs;
+use crate::piece::Board;
+
+fn king<'a>(pieces: &Board, strs: &Vec<char>, is_white: bool) -> Result<Board, Errs> {
+    // todo 動こうとしたマスに敵の駒の攻撃がないか
+    // todo キャスリング
+    todo!()
+}
