+++ src/movements/knight.rs
@@ -1,0 +1,7 @@
+use crate::errs::Errs;
+use crate::piece::Board;
+
+fn knight<'a>(pieces: &Board, strs: &Vec<char>, is_white: bool) -> Result<Board, Errs> {
+    // todo 動きたいマスに動くことができるか
+    todo!()
+}
