+++ src/movements/rook.rs
@@ -1,0 +1,9 @@
+use crate::errs::Errs;
+use crate::piece::Board;
+
+fn rook<'a>(pieces: &Board, strs: &Vec<char>, is_white: bool) -> Result<Board, Errs> {
+    // todo 動こうとしたマスが直線上にあるか
+    // todo 動こうとしたマスがなにかの駒で防がれていないか
+    // todo キャスリング
+    todo!()
+}
